finish json messages + error user bad

+++ app/Catalogs/Actions/CategoryAction.php
@@ -4,13 +4,19 @@ namespace App\Catalogs\Actions;
 
 use App\Base\Actions\Action;
 use App\Models\Category as Model;
+use App\Models\Help;
 use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Http\JsonResponse;
 
 class CategoryAction extends Action
 {
     private array $categories;
 
+    private array $response;
+
     private int $total;
+
+    private int $count;
 
     public function getAllPages(): Collection
     {
@@ -40,29 +46,45 @@ class CategoryAction extends Action
         return $this->item;
     }
 
-    public function store(array $request): bool
+    public function store(array $request): JsonResponse
     {
         Model::create($request);
-        Model::flushQueryCache();
+        $this->response = [
+            'message' => 'Категория успешно добавлена!',
+        ];
 
-        return true;
+        return response()->success($this->response);
     }
 
-    public function update(array $request, int $id): Model
+    public function update(array $request, int $id): JsonResponse
     {
         $this->item = Model::findOrFail($id);
         $this->item->update($request);
         Model::flushQueryCache();
+        $this->response = [
+            'message' => 'Категория успешно обновлена!',
+        ];
 
-        return $this->item;
+        return response()->success($this->response);
     }
 
-    public function delete(int $id): bool
+    public function delete(int $id): JsonResponse
     {
+        $this->count = Help::dontCache()->where('category_id', $id)->count();
+        if ($this->count > 0) {
+            $this->response = [
+                'message' => 'Категория не может быть удалена, так как не удалены все заявки связанные с ней!',
+            ];
+
+            return response()->error($this->response);
+        }
         $this->item = Model::findOrFail($id);
         $this->item->forceDelete();
         Model::flushQueryCache();
+        $this->response = [
+            'message' => 'Категория успешно удалена!',
+        ];
 
-        return true;
+        return response()->success($this->response);
     }
 }
