+++ app/Catalogs/Actions/StatusAction.php
@@ -3,14 +3,20 @@
 namespace App\Catalogs\Actions;
 
 use App\Base\Actions\Action;
+use App\Models\Help;
 use App\Models\Status as Model;
 use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Http\JsonResponse;
 
 class StatusAction extends Action
 {
     private array $statuses;
 
+    private array $response;
+
     private int $total;
+
+    private int $count;
 
     public function getAllPages(): Collection
     {
@@ -39,29 +45,48 @@ class StatusAction extends Action
         return $this->item;
     }
 
-    public function store(array $request): bool
+    public function store(array $request): JsonResponse
     {
-        Model::create($request);
+        $this->item = Model::create($request);
         Model::flushQueryCache();
+        $this->response = [
+            'message' => 'Статус успешно добавлен!',
+        ];
 
-        return true;
+        return response()->success($this->response);
+
     }
 
-    public function update(array $request, int $id): Model
+    public function update(array $request, int $id): JsonResponse
     {
         $this->item = Model::findOrFail($id);
         $this->item->update($request);
         Model::flushQueryCache();
+        $this->response = [
+            'message' => 'Статус успешно обновлён!',
+        ];
 
-        return $this->item;
+        return response()->success($this->response);
     }
 
-    public function delete(int $id): bool
+    public function delete(int $id): JsonResponse
     {
+        $this->count = Help::dontCache()->where('status_id', $id)->count();
+        if ($this->count > 0) {
+            $this->response = [
+                'message' => 'Статус не может быть удалён, так как не удалены все заявки связанные с ним!',
+            ];
+
+            return response()->error($this->response);
+        }
+
         $this->item = Model::findOrFail($id);
         $this->item->forceDelete();
         Model::flushQueryCache();
+        $this->response = [
+            'message' => 'Статус успешно удалён!',
+        ];
 
-        return true;
+        return response()->success($this->response);
     }
 }
